Convert use service to TRPC
