+++ src/api/services/router.ts
@@ -1,10 +1,14 @@
 import { t } from "./t";
-import { createUtilService } from "./util/service";
+import { UtilService } from "./util/service";
+import { UserService } from "./user/service";
 
-export function createApiRouter() {
-  return t.router({
-    util: createUtilService(),
-  });
+export function createApiRouter(services: ApiServices) {
+  return t.router(services);
 }
+
+export type ApiServices = {
+  util: UtilService;
+  user: UserService;
+};
 
 export type ApiRouter = ReturnType<typeof createApiRouter>;
